requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-api-python-client==2.149.0
requests==2.32.3
pandas==2.2.3
scikit-learn==1.5.2
plotly==5.24.1
python-dotenv==1.0.1
numpy==1.26.4